+++ db.json
@@ -10,7 +10,7 @@
       "address": "서울 강남구 강남대로 478, 1층 (논현동)",
       "latitude": 37.5051069,
       "longitude": 127.0246713,
-      "likes": 14,
+      "likes": 16,
       "comments": []
     },
     {
@@ -23,7 +23,7 @@
       "address": "서울 강남구 강남대로118번길 51, 103호 (논현동)",
       "latitude": 37.5196301,
       "longitude": 127.0381309,
-      "likes": 8,
+      "likes": 9,
       "comments": [
         {
           "id": 1726935145062,
